Release site

+++ CorallJewelry/Controllers/AdminController.cs
@@ -11,6 +11,9 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Extensions;
 using CorallJewelry.Controllers.Executors.Admin;
+using System.Security.Cryptography;
+using afc_studio.Models.Entitys;
+using System.Text.Json;
 
 namespace CorallJewelry.Controllers
 {
@@ -19,11 +22,12 @@ namespace CorallJewelry.Controllers
         #region Header
         private readonly ILogger<HomeController> _logger;
         private BackendContext db { get; set; }
+        private MainContext db2 { get; set; }
         public AdminController(ILogger<HomeController> logger)
         {
             db = new BackendContext(new DbContextOptions<BackendContext>());
             _logger = logger;
-
+            db2 = new MainContext(new DbContextOptions<MainContext>());
             if (!db.Users.Any(x => x.Login == "adminCoral" && x.Password == "adminCoralJewelry202056"))
             {
                 db.Users.Add(new Models.User() { Login = "adminCoral", Password = "adminCoralJewelry202056", Type = TypeUser.Admin, LastSession = DateTime.Now });
@@ -63,25 +67,42 @@ namespace CorallJewelry.Controllers
         }
         public IActionResult Chats()
         {
-            return View();
+            //if (!Auth())
+            //{
+            //    return View("Login");
+            //}
+            return View(AllExecutors.ChatExecutor.GetModel());
         }
         public IActionResult Contacts()
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             var cont = AllExecutors.ContactExecutor.GetContact();
             return View(cont);
         }
         public IActionResult Prices()
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             var price = AllExecutors.PriceExecutor.GetAllPriceLists();
             return View(price);
         }
         public IActionResult Requests()
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             var req = AllExecutors.RequestExecutor.GetRequest();
             return View(req);
         }
         public IActionResult Login()
         {
+
             return View();
         }
         #endregion
@@ -103,6 +124,10 @@ namespace CorallJewelry.Controllers
         [HttpPost]
         public IActionResult AddProduct(List<IFormFile> images, string name, string about, double price, string weight, string stone, string metall, string type)
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             AllExecutors.ProductsExecutor.AddProducts(images, name, about, price, weight, stone, metall, type);
             var model = AllExecutors.ProductsExecutor.GetProducts("all");
             return View("Products", model);
@@ -110,6 +135,10 @@ namespace CorallJewelry.Controllers
         [HttpPost]
         public IActionResult EditProducts(int id, List<IFormFile> images, string name, string about, double price, string weight, string stone, string metall, string type,string action)
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             if (action == "edit")
             {
 
@@ -128,6 +157,10 @@ namespace CorallJewelry.Controllers
         [HttpPost]
         public IActionResult EditContacts(string EmailInfo, string PhoneInfo, string VKLink, string OKLink, string InstagramLink, string addressT, string addressS)
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             AllExecutors.ContactExecutor.EditContact(EmailInfo,PhoneInfo,VKLink,OKLink,InstagramLink,addressT,addressS);
             var model = AllExecutors.ContactExecutor.GetContact();
             return View("Contacts", model);
@@ -138,12 +171,20 @@ namespace CorallJewelry.Controllers
         [HttpPost]
         public IActionResult AddPriceList(string namePrice)
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             AllExecutors.PriceExecutor.CreateList(namePrice);
             var model = AllExecutors.PriceExecutor.GetAllPriceLists();
             return View("Prices", model);
         }
         public IActionResult AddPrice(string name, string price, int idList)
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             AllExecutors.PriceExecutor.AddPrice(idList,name,price);
             var model = AllExecutors.PriceExecutor.GetAllPriceLists();
             return View("Prices", model);
@@ -152,6 +193,10 @@ namespace CorallJewelry.Controllers
         [HttpPost]
         public IActionResult EditPrices(string action, int idlist, string name, string price, int idprice)
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             if (action == "edit")
             {
                 AllExecutors.PriceExecutor.EditPrice(idprice, name, price);
@@ -169,9 +214,64 @@ namespace CorallJewelry.Controllers
         [HttpPost]
         public IActionResult DeleteReq(int id)
         {
+            if (!Auth())
+            {
+                return View("Login");
+            }
             AllExecutors.RequestExecutor.DeleteRequest(id);
             var model = AllExecutors.RequestExecutor.GetRequest();
             return View("Requests",model);
+        }
+        #endregion
+        #region ChatPost
+        [HttpPost]
+        public string GetDialog(string user)
+        {
+            afc_studio.Models.Objects.User my;
+            if (!db2.Users.Any(j => j.Name == user))
+            {
+                db2.Users.Add(new afc_studio.Models.Objects.User()
+                {
+                    Type = "unnamed",
+                    Name = user,
+                    Dialogs = new List<afc_studio.Models.Objects.Dialog>()
+                    {
+                        new afc_studio.Models.Objects.Dialog
+                        {
+                            Name = "dialog_" + user +  "_" + DateTime.Now.Day.ToString() +  DateTime.Now.Minute.ToString() + DateTime.Now.Second.ToString()
+                        }
+                    }
+                });
+                db2.SaveChanges();
+            }
+            my = db2.Users.Where(d => d.Name == user).Include(s => s.Dialogs).FirstOrDefault();
+            return (my.Dialogs[0].Name);
+        }
+
+        [HttpPost]
+        public string SendMessage(string text, string dialog, string user, string typeuser = "unnamed", int dialog_num = 0)
+        {
+            var mydialog = db2.Dialogs.Where(j => j.Name == dialog).Include(g => g.Messages).FirstOrDefault();
+            var myuser = "admin";
+
+            mydialog.Messages.Add(new afc_studio.Models.Objects.Message()
+            {
+                Text = text,
+                Time = DateTime.Now,
+                User = myuser
+            });
+            db2.SaveChanges();
+            return "1";
+        }
+        [HttpPost]
+        public string GetMessages(string dialog)
+        {
+            if (dialog != null)
+            {
+                var mydialog = db2.Dialogs.Where(j => j.Name == dialog).Include(g => g.Messages).FirstOrDefault();
+                return JsonSerializer.Serialize(mydialog.Messages);
+            }
+            return null;
         }
         #endregion
